Switching to hashtable index
There could be something wrong with Mnemosyne
or the RBtree or the interaction between them.

Or may be if i run hashtable long enough, I'll face
the same trouble. IDK. But at least switching to the
hashtable index has made life easier when the program
actually terminates succesfully.

It is impossible to debug RBTree as GDB doesn't produce a
back trace.