Fixed transaction error handling
When a Tx begins, it saves processor state
in tx->jb (_ITM_begin_transaction). GLIBC setjmp
performs the same function but cannot be used as
the state saved by it is invalidated when its caller
returns. And _ITM_begin_transaction returns. Hence,
the assembly hack.

This state is used to restore state and restart execution
when a Tx fails. Again, this function can be performed by
GLIBC longjmp but only if the corresponding state is stored
by setjmp. _ITM_siglongjmp performs this function.

So long as state is properly saved and restored, everything
works fine. This wasn't happening.

1. State wasn't being saved in MTM_begin_transaction.
2. Even when it was saved, it was being restored in _ITM_siglongjmp.
* _ITM_siglongjmp was using its second parameter stored in %rsi,
when infact it should be using %rdi - the first parameter.

Changed the Tx descr to use jmp_buf. Stack pointer is the first 8 bytes
of this buffer. Changed Mnemosyne to use _ITM_siglongjmp although
mtm_longjmp can also be used.